// asb-good-foods-user/src/app/common/models/review.model.ts
export class Review {
    userId: string;
    reviewDocId: string;
    ratings: number;
    description: string;
    title: string;
    imageUrl: string;
    date: any;
    reviewStatus: string;
    imageStatus: string;
    nameStatus: string;
    productId: string;
    userName: string;
    productImg: string;
    reviewId:number;
    pilled:[];
    unpilled:[];
}